Movie Site (Entery Level)
I worked on the design aspect of this movie website project by focusing solely on visual aesthetics. Those who wish to can develop the Back-End of the project and create a dynamic website. I saved the data dynamically only in the "movie.json" database. I listed and displayed this data in various fields within "app.js". Additionally, I activated the search input using only "json". (There is no MP4 player or background player for movies on the site, as in The Boys. They are merely placed to enhance visual richness.) If I have time in the future, I can update the site to a dynamic, admin-panelled state. Stay tuned. :)